requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
sentence-transformers>=2.5.0
openai>=1.10.0
